src/prediction_instance/unet_upload_cloud_func/requirements.txt
affine==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
cachetools==5.3.2
certifi==2024.2.2
cftime==1.6.3
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
decorator==5.1.1
filelock==3.13.1
fiona==1.9.5
frozenlist==1.4.1
fsspec==2024.2.0
gcsfs==2024.2.0
geopandas==0.14.3
google-api-core==2.16.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
h5netcdf==1.3.0
h5py==3.10.0
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
netCDF4==1.6.5
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
packaging==23.2
pandas==2.2.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
pytz==2024.1
rasterio==1.3.9
requests==2.31.0
requests-oauthlib==1.3.1
rioxarray==0.15.1
rsa==4.9
scikit-learn==1.4.0
scipy==1.12.0
shapely==2.0.2
six==1.16.0
snuggs==1.4.7
sympy==1.12
threadpoolctl==3.2.0
torch==2.2.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
xarray==2024.1.1
yarl==1.9.4
